Refactor release script
